Zipl VirtIO bootloader code
This is a patch to enable booting inside KVM with virtio hard drives.

It creates a first stage bootloader (virtio1) that reads sector descriptors
from offset 0x20 and following of the MBR. The resulting binary should be
put in the pc-bios directory of qemu.

It also adds a second stage bootloader (virtio2) that is written to the
disk by zipl. This second stage bootloader shows the boot menu and loads
the kernel and initrd.

The second stage bootloader is only written to the disk if virtio is detected.

Signed-off-by: Alexander Graf <agraf@suse.de>