zipl: Use yield when polling for virtio requests to be done
We're polling to wait for virtio to process its queue. Now, we currently
do this without passing control to the host for more than the kick, which
means that on UP hosts we end up slowing the actual transfer down.

With this patch, the host gets a hint that the guest isn't really doing
anything useful.

Signed-off-by: Alexander Graf <agraf@suse.de>