zipl: virtio: send block requests of full size
So far, we've been splitting block requests into 512 byte chunks and
waited for each 512 byte request to finish before kicking off the next.

Obviously, this was slow.

With this patch, we simply issue the full contiguous length we get from
the loading table, reducing the time spent loading the kernel/initrd to
almost nothing.

Signed-off-by: Alexander Graf <agraf@suse.de>